upselling mobile, obtener monto meta y guardar regalo

+++ app/Portal.Consultoras.Web/Models/RegaloOfertaFinalModel.cs
@@ -9,7 +9,7 @@ namespace Portal.Consultoras.Web.Models
 
         public int CampaniaId { get; set; }
 
-        public int ConsultoraId { get; set; }
+        public long ConsultoraId { get; set; }
 
         public decimal MontoPedido { get; set; }
